% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSB_RNs.R
\name{CSB_RNs}
\alias{CSB_RNs}
\title{Compute the function of CSB_RNs}
\usage{
CSB_RNs(data, netParameterMatrix)
}
\arguments{
\item{data}{profileR,The data type provided in the package}

\item{netParameterCsvfile}{The exported CSV network file from
the STRING database is entered into cytoscape, and the exported
file is calculated by the plug-in networkAnalysis}
}
\value{
data.frame Returns the calculated result data
}
\description{
The CSB_RNs parameters were calculated according to
the expression spectrum and network topology parameters.
}
\seealso{
\code{\link{profileR}}
}
\author{
Xingyi Liu
}
