% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneentropyfilter.R
\name{geneentropyfilter}
\alias{geneentropyfilter}
\title{A function that USES entropy to clean data}
\usage{
geneentropyfilter(
  radduoData,
  filtercutoff = 0.1,
  na.dealed = TRUE,
  all.zero.remove = TRUE
)
}
\arguments{
\item{radduoData}{data.frame,The row name is the gene and the column name is the sample (data.frame).}

\item{filtercutoff}{numeric,The default value is 0.1,The entropy value less than this parameter is removed.}

\item{na.dealed}{logical,The default value is TRUE,Whether to remove all 0 rows.}

\item{all.zero.remove}{logical,The default value is TRUE,Whether to remove all NA rows.}
}
\value{
data.frame,The gene expression profile after removing the low entropy value
}
\description{
The rows of gene expression profiles with low entropy,
all 0 and null values were deleted.Give the gene expression
profile (data.frame) and return the cleaned data (data.frame).
}
\author{
Xingyi Liu
}
