% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloader.R
\docType{function}
\name{get.GEO.data}
\alias{get.GEO.data}
\alias{dowloadGEO}
\title{get.GEO.data function}
\usage{
get.GEO.data(GEO.acession = "28735", platfrom = "GPL6244")
}
\arguments{
\item{GEO.acession}{GEO acession id to dowload file}

\item{raw.data}{if TRUE,dowload raw data}

\item{result.data}{gene profiler data}
}
\description{
dowload gene profiler from ncbi ftp website and return data.frame about that
}
\author{
Xingyi Liu
}
