% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOtermsbarplot.R
\name{plotGOtermsbarplot}
\alias{plotGOtermsbarplot}
\title{Draw a bar chart for GO enrichment analysis}
\usage{
plotGOtermsbarplot(
  geneSet,
  label.min.n.words = 4,
  label.min.n.char = 20,
  display.number = 10,
  Species.label = "hsa",
  enrich.pvalue = 0.05,
  plot.type = 2,
  coord_flip = TRUE,
  CPCOLS = c("#2f5688", "#0F1F95", "#CC0000"),
  xlab = "GO term",
  ylab = "-log10(FDR)",
  title = "GO enrichment plot",
  face = "bold",
  axis.text.color = "gray40",
  axis.text.font.size = 12,
  axis.title.font.size = 14,
  legend.text.font.size = 10
)
}
\arguments{
\item{geneSet}{as.vector,gene dataset.}
}
\description{
According to the input geneSymol collection,
 convert to entrazid and draw three types of bar charts.
}
\author{
Xingyi Liu
}
