% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.matrix.rn.smooth.R
\name{Predict.matrix.rn.smooth}
\alias{Predict.matrix.rn.smooth}
\title{River Network Smooths}
\usage{
\method{Predict.matrix}{rn.smooth}(object, data)
}
\arguments{
\item{object}{For the smooth.construct method a smooth specification object,
usually generated by a term s(x,...,bs="rn", xt=list(polys=foo)). x is a
factor variable giving labels for geographic districts, and the xt argument
is obligatory: see details. For the Predict.Matrix method an object of class
"rn.smooth" produced by the smooth.construct method.}

\item{data}{a list containing just the data (including any by variable)
required by this term, with names corresponding to object$term
(and object$by). The by variable is the last element.}
}
\value{
An object of class "rn.smooth" or a matrix mapping the coefficients of the
river network smooth to the predictions for the nodes listed in data.
}
\description{
For data observed over river networks, a simple river network smoother can be
appropriate. These functions provide such a smoother class for mgcv.
}
