% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.construct.rn.smooth.spec.R
\name{smooth.construct.rn.smooth.spec}
\alias{smooth.construct.rn.smooth.spec}
\title{River Network Smoothers in GAMs}
\usage{
smooth.construct.rn.smooth.spec(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, usually generated by a term 
s(...,bs="cr",...), s(...,bs="cs",...) or s(...,bs="cc",...).}

\item{data}{a list containing just the data (including any by variable) 
required by this term, with names corresponding to object$term 
(and object$by). The by variable is the last element.}

\item{knots}{a list containing any knots supplied for basis setup — in same 
order and with same names as data. Can be NULL. See details.}
}
\value{
An object of class "rn.smooth". In addition to the usual elements of a smooth 
class documented under smooth.construct, this object will contain:

xp	
giving the knot locations used to generate the basis.
}
\description{
Description here.
}
\details{
The constructor is not normally called directly, but is rather used 
internally by gam. To use for basis setup it is recommended to use 
\link{smooth.construct2}.
}
\examples{

}
\references{

}
