% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Draw a volcano plot}
\usage{
volcanoPlot(object, xlab = "", ylab = "", title = "", points = 10)
}
\arguments{
\item{object}{profileR,The data type is getted from analysisIt function.}

\item{xlab}{character,Notes on the horizontal axis of the picture.}

\item{ylab}{character,Note the vertical axis of the drawing.}

\item{title}{character,The main title of the picture.}

\item{points}{numeric,the default is 10,The number of store labels up and down.}
}
\description{
Based on the results of the analysisIt function, a volcano map is drawn
}
\author{
Xingyi Liu
}
